requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.1
Flask-Migrate==4.0.7
python-dotenv==1.0.1
Pillow>=11.0.0
requests>=2.31.0
reportlab>=4.0.0
Markdown==3.7
pint>=0.24
